criando endipoints de get movies com paginação e get movie by id com tratamento de erro

+++ catalog/src/main/java/com/henrique/catalog/infra/padronize/SuccessListDataResponse.java
@@ -6,12 +6,12 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 public record SuccessListDataResponse(List<?> data,
-                                      Long page,
-                                      Long pageSize,
+                                      Integer page,
+                                      Integer pageSize,
                                       Long totalElements,
                                       @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
                                       LocalDateTime timestamp) {
-    public SuccessListDataResponse(List<?> data, Long page, Long pageSize, Long totalElements) {
+    public SuccessListDataResponse(List<?> data, Integer page, Integer pageSize, Long totalElements) {
         this(data,
                 page,
                 pageSize,
